+++ Clip Hold/StandardPhrase/StandardPhraseWindowView.swift
@@ -244,19 +244,9 @@ struct StandardPhraseWindowView: View {
         }
     }
     
-    
-    
     var body: some View {
         ZStack { // ZStackでコンテンツとメッセージを重ねる
-            if #available(macOS 26, *) {
-                Color.clear
-                    .glassEffect(in: .rect(cornerRadius: 16.0))
-                    .overlay(colorScheme == .dark ? Color.black.opacity(0.2) : Color.white.opacity(0.5))
-                    .ignoresSafeArea()
-            } else {
-                VisualEffectView(material: .menu, blendingMode: .behindWindow)
-                    .ignoresSafeArea()
-            }
+            SharedWindowBackground()
             
             ZStack { // メインコンテンツを囲むZStack
                 VStack(spacing: 0) {